requirements.txt
requests==2.22.0
lxml==4.3.4
pymongo==3.8.0
PyYAML==5.1.1
